feat(api,cli): float64 support in GeneReconModel; --dtype now honored for reconcile/eval
Add a dtype kwarg to GeneReconModel (default float32, back-compat) threaded into theta,
receiver_weights, and origination_weights; the forward compute already follows theta's
dtype and upcasts internally, so no kernel changes are needed. Wire the CLI's --dtype
into build_model and the fidelity bench's reconcile_at_alerax_rates, both of which
previously computed the dtype but silently dropped it. Correct docs/cli.md accordingly,
and add tests proving the dtype flows end-to-end and that omitting it stays float32.

+++ gpurec/bench/fidelity.py
@@ -53,7 +53,7 @@ def reconcile_at_alerax_rates(species, genes, rates, *, device="cuda", dtype=Non
     D, L, T = float(rates[0]), float(rates[1]), float(rates[2])
     gene_list = genes if isinstance(genes, list) else [genes]
     model = GeneReconModel(species, gene_list, mode=mode, device=device,
-                           solver_options=solver_options)
+                           dtype=dtype, solver_options=solver_options)
     # gpurec theta order is [log2 D, log2 L, log2 T] == AleRax (D, L, T) column order
     triple = torch.tensor([math.log2(D), math.log2(L), math.log2(T)],
                           dtype=model.theta.dtype, device=model.theta.device)
